refactor: revamp architecture, optimize ingestion, and simplify admin commands
Architecture Changes:
- Replaced the continuous background worker with a systemd timer-driven nightly maintenance task (maintenance_main.py, quote-finder-maintenance.timer).
- Renamed main.py to bot_main.py to clarify entry points.
- Replaced legacy systemd files with quote-finder-bot.service and quote-finder-maintenance.service.

Ingestion Optimization:
- Implemented smart delta-updates for the refresh flow (refresh_fic):
  - Skips EPUB downloads entirely if the FicHub metadata hash matches the database version.
  - Compares chapter hashes to detect modifications, rewriting and embedding only new or edited chapters to save CPU and Qdrant resources.
- Updated ChunkBuilder to use chapter-scoped chunk numbering to avoid sequence collisions on delta updates.
- Added delete_chapters methods to Qdrant store and repository to cleanly handle partial vector replacements.

Admin & UI Adjustments:
- Removed complex slash-command modals (ui/admin_views.py) in favor of direct text prefix commands (!qf ingest, !qf rebuild, !qf refresh, !qf connect, !qf status).
- Integrated dynamic background job tracking directly into discord command messages.
- Added pagination support for !qf status embeds to handle large lists of fics.

Database & Fixes:
- Added Alembic migration to change jobs.guild_id to BigInteger, fixing PostgreSQL DataError on large Discord IDs.
- Fixed SQLAlchemy 2.0+ selectinload errors by explicitly mapping the FicGuild relationship on the Fic model.
- Fixed timestamps to properly record last_refreshed_at on successful ingestions.

Documentation:
- Updated README.md and docs/ai/CURRENT_STATE.md to reflect the new systemd deployment instructions and architecture layout.

+++ database/models/job.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime, timezone
-from sqlalchemy import String, Integer, DateTime, ForeignKey
+from sqlalchemy import String, Integer, BigInteger, DateTime, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 from database.models.base import Base
 
@@ -26,4 +26,4 @@ class Job(Base):
     completed_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=True)
 
     target_url: Mapped[str] = mapped_column(String, nullable=True)
-    guild_id: Mapped[int] = mapped_column(Integer, nullable=True) 
+    guild_id: Mapped[int] = mapped_column(BigInteger, nullable=True)
